ultimate admin operation

+++ Reviook_app/src/main/java/it/unipi/dii/reviook_app/controllers/AdminController.java
@@ -126,6 +126,7 @@ public class AdminController {
     @FXML
     void logsAction() {
         clearList();
+        deleteElemButton.setDisable(true);
         unReportButton.setDisable(true);
         ArrayList<Log> list = adminManager.loadLogs();
         obsListLog.addAll(list);
@@ -142,10 +143,22 @@ public class AdminController {
                     if (selectedCell == null) {
                         return;
                     }
-
-                    adminManager.restoreLog(selectedCell);
-                    adminManager.deleteLog(selectedCell);
-                    obsListLog.remove(selectedCell);
+                    if (selectedCell.getOperation().equals("delete")) {
+                        if (adminManager.restoreLog(selectedCell)) {
+                            // reported element restored with success
+                            if (adminManager.deleteLog(selectedCell)) {
+                                // selected log deleted with success
+                                obsListLog.remove(selectedCell);
+                            } else {
+                                // can't delete selected log --- retry delete log
+                                adminManager.deleteLog(selectedCell);
+                                actionTarget.setText("Error: unable to restore selected element");
+                            }
+                        } else {
+                            // can't restore reported elem
+                            actionTarget.setText("Error: unable to restore selected element");
+                        }
+                    }
                 }
             }
         });
@@ -165,7 +178,6 @@ public class AdminController {
                 };
             }
         });
-        //TODO fare sul doppio click destro un ripristino dei log
     }
 
     @FXML
@@ -193,6 +205,9 @@ public class AdminController {
         actionTarget.setText("");
         if (bookOption.isSelected()) {
             Report selectedBook = (Report) bookList.getSelectionModel().getSelectedItem();
+            if (selectedBook == null) {
+                return;
+            }
             Book bookForBackup = bookManager.getBookByID(selectedBook.getBook_id());
             if (bookManager.deleteBook(selectedBook.getBook_id())) {
                 // book deleted with success
@@ -203,25 +218,13 @@ public class AdminController {
                     resetRightDetail();
                 } else {
                     // can't delete report -- reinsert book
-                    DateTimeFormatter formatter = DateTimeFormatter.ofPattern("d/MM/yyyy");
-                    String stringDate = bookForBackup.getPublication_day().toString()
-                            + "/" +
-                            bookForBackup.getPublication_month().toString()
-                            + "/" +
-                            bookForBackup.getPublication_year().toString();
-                    LocalDate date = LocalDate.parse(stringDate, formatter);
-                    bookManager.addBook(
-                            bookForBackup.getNum_pages(),
-                            bookForBackup.getImage_url(),
-                            bookForBackup.getLanguage_code(),
-                            date,
-                            bookForBackup.getBook_id(),
-                            bookForBackup.getTitle(),
-                            bookForBackup.getIsbn(),
-                            bookForBackup.getDescription(),
-                            bookForBackup.getGenres(),
-                            bookForBackup.getAuthors()
-                    );
+                    if (bookManager.addBookMongo(bookForBackup)) {
+                        // book restored in mongoDB with success
+                        bookManager.addBookN4J(bookForBackup);
+                    } else {
+                        // can't restore book
+                        actionTarget.setText("Error: unable to remove Book");
+                    }
                     actionTarget.setText("Error: unable to remove Book");
                 }
             } else {
@@ -229,28 +232,49 @@ public class AdminController {
             }
         } else if (userOption.isSelected()) {
             User selectedUser = (User) usersList.getSelectionModel().getSelectedItem();
+            if (selectedUser == null) {
+                return;
+            }
             if (userManager.deleteUserMongo(selectedUser, "user")) {
-                if (!userManager.deleteUserN4J(selectedUser, "user"))
+                // user delete with success from MongoDB
+                if (userManager.deleteUserN4J(selectedUser, "user")) {
+                    // user delete with success from N4J
+                    obsUserList.remove(selectedUser);
+                    addCustomFactory("user");
+                    resetRightDetail();
+                } else {
+                    // can't delete user from N4J
                     actionTarget.setText("Error: unable to delete user");
+                }
             } else {
+                // can't delete user form MongoDB
                 actionTarget.setText("Error: unable to delete user");
             }
-            obsUserList.remove(selectedUser);
-            addCustomFactory("user");
-            resetRightDetail();
         } else if (authorOption.isSelected()) {
             Author selectedAuthor = (Author) authorsList.getSelectionModel().getSelectedItem();
+            if (selectedAuthor == null) {
+                return;
+            }
             if (userManager.deleteUserMongo(selectedAuthor, "author")) {
-                if (!userManager.deleteUserN4J(selectedAuthor, "author"))
+                // author delete with success from MongoDB
+                if (userManager.deleteUserN4J(selectedAuthor, "author")) {
+                    // author delete with success from N4J
+                    obsAuthorList.remove(selectedAuthor);
+                    addCustomFactory("author");
+                    resetRightDetail();
+                } else {
+                    // can't delete author from N4J
                     actionTarget.setText("Error: unable to delete author");
+                }
             } else {
+                // can't delete author form MongoDB
                 actionTarget.setText("Error: unable to delete author");
             }
-            obsAuthorList.remove(selectedAuthor);
-            addCustomFactory("author");
-            resetRightDetail();
         } else if (reviewOption.isSelected()) {
             Report selectedReview = (Report) reviewList.getSelectionModel().getSelectedItem();
+            if (selectedReview == null) {
+                return;
+            }
             if (bookManager.deleteReview(selectedReview.getReview_id(), selectedReview.getBook_id())) {
                 // review removed with success
                 if (adminManager.deleteReport(selectedReview, false)) {
@@ -284,25 +308,33 @@ public class AdminController {
     void unReport() {
         if (bookOption.isSelected()) {
             Report selectedBook = (Report) bookList.getSelectionModel().getSelectedItem();
-            int index = bookList.getSelectionModel().getSelectedIndex();
-            if (!adminManager.deleteReport(selectedBook, true))
-                actionTarget.setText("Error: unable to remove bookReport");
-            obsBooksList.remove(selectedBook);
-            addCustomFactory("book");
-            resetRightDetail();
+            if (adminManager.deleteReport(selectedBook, true)) {
+                // reported book deleted with success
+                obsBooksList.remove(selectedBook);
+                addCustomFactory("book");
+                resetRightDetail();
+            } else {
+                // can't delete reported book
+                actionTarget.setText("Error: unable to remove reported book");
+            }
         } else if (reviewOption.isSelected()) {
             Report selectedReview = (Report) reviewList.getSelectionModel().getSelectedItem();
-            if (!adminManager.deleteReport(selectedReview, true))
+            if (adminManager.deleteReport(selectedReview, true)) {
+                // reported review deleted with success
+                obsListReview.remove(selectedReview);
+                addCustomFactory("review");
+                resetRightDetail();
+            } else {
+                // can't delete reported review
                 actionTarget.setText("Error: unable to remove reviewReport");
-            obsListReview.remove(selectedReview);
-            addCustomFactory("review");
-            resetRightDetail();
+            }
         }
     }
 
     @FXML
     void searchAction() {
         clearList();
+        deleteElemButton.setDisable(false);
         if (bookOption.isSelected()) {
             ArrayList<Report> list = adminManager.loadBookReported();
             obsBooksList.addAll(list);
